django-on-docker/app/requirements.txt
Django==4.2.9
psycopg2-binary==2.9.6
djangorestframework==3.14.0
djangorestframework-simplejwt
djangochannelsrestframework==0.3.0
channels-redis==3.3.1
django-cors-headers
django_extensions
werkzeug
pyopenssl
channels[daphne]